+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Bumped all plugins / dependencies to latest versions
 * Moved all `slick2d-core` tests in `src/test/` and added test data. Test classes are not shipped any more in the main JAR
 * Updated `Renderable` to include more `draw()` methods common to all implementations, thanks to [lucas_cimon](https://bitbucket.org/kevglass/slick/pull-request/15/making-renderable-class-more-polymorphic/diff)
+* Fixed `RoundedRectangle.contains()` after calling `setWidth()` / `setHeight()` thanks to [Joshua Hertlein](https://bitbucket.org/kevglass/slick/issue/41/roundedrectangles-setwidth-and-setheight)
 
 ## v1.0.0 (2014-04-16)